Crowbar 0.36 - 2016-11-03_0.36
What's New

=== Unpack ===
New Unpack tab. Unpack and extract files from one or more VPK files.

=== Decompile ===
Decompile MDL version 36 (only reference mesh, physics mesh, and animation SMD files; QC file is incomplete).
Decompile MDL version 37 (only reference mesh, physics mesh, and animation SMD files; QC file is incomplete).
Decompile MDL version 47 as if it were version 46.
FIX: Using "$DeclareSequence QCI files" decompile option fails to complete for models without texture file names, which is common for animation-only MDL files.
FIX: Decompiling fails to complete for models without texture file names, which is common for animation-only MDL files.
FIX: Decompiling of a model that has bone names with spaces causes Crowbar to write incorrect $CollisionJoints block of lines and to write $CollisionText lines that should not be there in QC file.

=== Compile ===
FIX: After selecting a QC file via Browse button, the command-line text box is not updated with the new QC file name.
FIX: Missing compiled files when Compile to Output Folder Full Path is set to the default that the model compiler uses.

=== Preview and View ===
FIX: Pasting in a file name that is surrounded by double-quotes prevents using anything else in the window.
FIX: When opening Crowbar that has a model in View and a model in Preview with different model versions, one of them would show the incorrect model version.

+++ Crowbar/Core/Compiler/Compiler.vb
@@ -433,10 +433,12 @@ Public Class Compiler
 			Try
 				If TheApp.Settings.CompileOutputFolderIsChecked Then
 					targetPathFileName = Path.Combine(Me.theModelOutputPath, Path.GetFileName(sourcePathFileName))
-					If File.Exists(targetPathFileName) Then
-						File.Delete(targetPathFileName)
+					If String.Compare(sourcePathFileName, targetPathFileName, True) <> 0 Then
+						If File.Exists(targetPathFileName) Then
+							File.Delete(targetPathFileName)
+						End If
+						File.Move(sourcePathFileName, targetPathFileName)
 					End If
-					File.Move(sourcePathFileName, targetPathFileName)
 				Else
 					targetPathFileName = sourcePathFileName
 				End If
@@ -446,6 +448,7 @@ Public Class Compiler
 				End If
 			Catch ex As Exception
 				'TODO: Write a warning message.
+				Dim debug As Integer = 4242
 			End Try
 		Next
 	End Sub
@@ -508,7 +511,8 @@ Public Class Compiler
 				myProcess.CancelOutputRead()
 				myProcess.CancelErrorRead()
 				myProcess.Kill()
-			Catch
+			Catch ex As Exception
+				Dim debug As Integer = 4242
 			End Try
 		End If
